Split out invocation from object.c/.h and veneer.c/.h
