+++ src/builtin/veneer.c
@@ -261,85 +261,6 @@ MORPHO_METHOD(MORPHO_CLONE_METHOD, Object_clone, BUILTIN_FLAGSEMPTY)
 MORPHO_ENDCLASS
 
 /* **********************************************************************
- * Invocation
- * ********************************************************************** */
-
-/** Creates a new invocation object */
-value invocation_constructor(vm *v, int nargs, value *args) {
-    value out=MORPHO_NIL;
-
-    if (nargs==2) {
-        value receiver = MORPHO_GETARG(args, 0);
-        value selector = MORPHO_GETARG(args, 1);
-        
-        if (!MORPHO_ISOBJECT(receiver) || !MORPHO_ISSTRING(selector)) {
-            morpho_runtimeerror(v, INVOCATION_ARGS);
-            return MORPHO_NIL;
-        }
-        
-        value method = MORPHO_NIL;
-        
-        objectclass *klass=object_getclass(receiver);
-        
-        if (dictionary_get(&klass->methods, selector, &method)) {
-            objectinvocation *new = object_newinvocation(receiver, method);
-
-            if (new) {
-                out = MORPHO_OBJECT(new);
-                morpho_bindobjects(v, 1, &out);
-            } else morpho_runtimeerror(v, ERROR_ALLOCATIONFAILED);
-        }
-        
-    } else morpho_runtimeerror(v, INVOCATION_ARGS);
-
-    return out;
-}
-
-/** Converts to a string for string interpolation */
-value Invocation_tostring(vm *v, int nargs, value *args) {
-    objectinvocation *inv=MORPHO_GETINVOCATION(MORPHO_SELF(args));
-    value out = MORPHO_NIL;
-
-    varray_char buffer;
-    varray_charinit(&buffer);
-    
-    morpho_printtobuffer(v, inv->receiver, &buffer);
-    varray_charwrite(&buffer, '.');
-    morpho_printtobuffer(v, inv->method, &buffer);
-    
-    out = object_stringfromvarraychar(&buffer);
-    if (MORPHO_ISSTRING(out)) {
-        morpho_bindobjects(v, 1, &out);
-    }
-    varray_charclear(&buffer);
-
-    return out;
-}
-
-/** Clones a range */
-value Invocation_clone(vm *v, int nargs, value *args) {
-    value self = MORPHO_SELF(args);
-    value out = MORPHO_NIL;
-    if (!MORPHO_ISINVOCATION(self)) return MORPHO_NIL;
-    
-    objectinvocation *slf = MORPHO_GETINVOCATION(self);
-    objectinvocation *new = object_newinvocation(slf->receiver, slf->method);
-    
-    if (new) {
-        out = MORPHO_OBJECT(new);
-        morpho_bindobjects(v, 1, &out);
-    } else morpho_runtimeerror(v, ERROR_ALLOCATIONFAILED);
-    
-    return out;
-}
-
-MORPHO_BEGINCLASS(Invocation)
-MORPHO_METHOD(MORPHO_PRINT_METHOD, Object_print, BUILTIN_FLAGSEMPTY),
-MORPHO_METHOD(MORPHO_TOSTRING_METHOD, Invocation_tostring, BUILTIN_FLAGSEMPTY),
-MORPHO_METHOD(MORPHO_CLONE_METHOD, Invocation_clone, BUILTIN_FLAGSEMPTY)
-MORPHO_ENDCLASS
-
-/* **********************************************************************
  * Error
  * ********************************************************************** */
 
@@ -404,11 +325,6 @@ void veneer_initialize(void) {
     /* Object */
     value objclass=builtin_addclass(OBJECT_CLASSNAME, MORPHO_GETCLASSDEFINITION(Object), MORPHO_NIL);
     morpho_setbaseclass(objclass);
-    
-    /* Invocation */
-    builtin_addfunction(INVOCATION_CLASSNAME, invocation_constructor, BUILTIN_FLAGSEMPTY);
-    value invocationclass=builtin_addclass(INVOCATION_CLASSNAME, MORPHO_GETCLASSDEFINITION(Invocation), objclass);
-    object_setveneerclass(OBJECT_INVOCATION, invocationclass);
     
     /* Error */
     builtin_addclass(ERROR_CLASSNAME, MORPHO_GETCLASSDEFINITION(Error), objclass);
