+++ src/classes/CMakeLists.txt
@@ -1,13 +1,14 @@
 target_sources(morpho
     PRIVATE
-        array.c    array.h
-        closure.c  closure.h
-        cmplx.c    cmplx.h
-        dict.c     dict.h
-        file.c     file.h
-        function.c function.h
-        list.c     list.h
-        range.c    range.h
-        strng.c    strng.h
-        system.c   system.h
+        array.c      array.h
+        closure.c    closure.h
+        cmplx.c      cmplx.h
+        dict.c       dict.h
+        file.c       file.h
+        function.c   function.h
+        invocation.c invocation.h
+        list.c       list.h
+        range.c      range.h
+        strng.c      strng.h
+        system.c     system.h
 )
